Final commit
Added QoL features including extra assets such as chairlifts and snow cannons, also added scoring system with a camera that showcases the ragdoll physics

+++ Assets/Scripts/Movement.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class Movement : MonoBehaviour
 {
-    public float speed;
-    public float acceleration;
-    public float timeTillAccel;
-    public float intervalBetweenAccel;
+    public int speed;
+    public int acceleration;
+    public int timeTillAccel;
+    public int intervalBetweenAccel;
     // Start is called before the first frame update
     void increaseSpeed()
     {
